+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+### Changed
+
+- For logged in users not having a username, the webID is displayed (#133).
+
+### Fixed
+
 ## [1.2.1] - 2024-06-17
 
 ### Added