+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Forced CSS's to not return content type application/ld+json, which induced a CORS error on some CSS server versions (#131).
+
 ## [1.2.1] - 2024-06-17
 
 ### Added